Wrap PublicCommentPage with ContentContainer and add appInsights to ContentContainer
